requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
certifi==2025.1.31
distro==1.9.0
docopt==0.6.2
docxcompose==1.4.0
docxtpl==0.19.1
frozenlist==1.6.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
lxml==5.3.2
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.4.3
num2words==0.5.14
numpy==2.2.5
openai==1.75.0
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
PyMuPDF==1.25.5
python-docx==1.1.2
python-dotenv==1.1.0
setuptools==79.0.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
yarl==1.20.0
